android/node.py
"""
Android Node - CH8 Agent node optimized for Android devices

Supports:
- Local models (llama.cpp via Termux)
- Cloud APIs (OpenAI, Claude, Groq, etc)
- Hybrid mode (local + cloud)
- Battery optimization
- Background service
"""

import asyncio
import aiohttp
import yaml
import psutil
import logging
from pathlib import Path
from typing import Dict, Any, Optional
from dataclasses import dataclass
from enum import Enum

# Setup logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger('ch8.android')


class OperationMode(Enum):
    """Node operation modes"""
    LOCAL = "local"      # Only local models
    CLOUD = "cloud"      # Only cloud APIs
    HYBRID = "hybrid"    # Both (smart routing)


@dataclass
class BatteryStatus:
    """Battery status information"""
    level: int           # 0-100
    is_charging: bool
    temperature: float   # Celsius


class AndroidNode:
    """
    CH8 Agent node optimized for Android

    Features:
    - Battery-aware processing
    - Temperature monitoring
    - Adaptive performance
    - Background service support
    """

    def __init__(self, config_path: str = None):
        self.config = self._load_config(config_path)
        self.mode = OperationMode(self.config.get('mode', 'hybrid'))
        self.is_running = False

        # Initialize based on mode
        self.local_model = None
        self.cloud_client = None

        logger.info(f"Initialized AndroidNode in {self.mode.value} mode")

    def _load_config(self, config_path: Optional[str]) -> Dict[str, Any]:
        """Load configuration"""
        if config_path is None:
            config_path = Path.home() / '.ch8' / 'config' / 'android-node.yaml'

        if not Path(config_path).exists():
            # Default configuration
            return {
                'mode': 'hybrid',
                'node': {
                    'id': 'android-node',
                    'platform': 'android'
                },
                'battery': {
                    'optimization': True,
                    'power_mode': 'efficient',
                    'max_cpu_usage': 50,
                    'temperature_limit': 40,
                    'min_battery_level': 20
                },
                'cloud': {
                    'provider': 'groq',
                    'model': 'llama3-8b-8192'
                }
            }

        with open(config_path, 'r') as f:
            return yaml.safe_load(f)

    async def start(self):
        """Start the Android node"""
        logger.info("Starting Android node...")

        # Initialize services based on mode
        if self.mode in [OperationMode.LOCAL, OperationMode.HYBRID]:
            await self._init_local_model()

        if self.mode in [OperationMode.CLOUD, OperationMode.HYBRID]:
            await self._init_cloud_client()

        self.is_running = True

        # Start main loop
        await self._main_loop()

    async def stop(self):
        """Stop the Android node"""
        logger.info("Stopping Android node...")
        self.is_running = False

    async def _init_local_model(self):
        """Initialize local model"""
        logger.info("Initializing local model...")

        local_config = self.config.get('local', {})
        model_path = local_config.get('model_path')

        if not model_path or not Path(model_path).exists():
            logger.warning("Local model not found, will use cloud fallback")
            return

        # Initialize llama.cpp model
        # (simplified - actual implementation would use ctypes/subprocess)
        logger.info(f"Loaded local model: {model_path}")
        self.local_model = {'path': model_path, 'ready': True}

    async def _init_cloud_client(self):
        """Initialize cloud API client"""
        logger.info("Initializing cloud client...")

        cloud_config = self.config.get('cloud', {})
        provider = cloud_config.get('provider', 'groq')

        self.cloud_client = {
            'provider': provider,
            'model': cloud_config.get('model'),
            'api_key': cloud_config.get('api_key'),
            'base_url': cloud_config.get('base_url')
        }

        logger.info(f"Cloud client ready: {provider}")

    async def _main_loop(self):
        """Main processing loop"""
        logger.info("Entering main loop...")

        while self.is_running:
            try:
                # Check battery status
                battery = self._get_battery_status()

                # Decide if we should process tasks
                if not self._should_process(battery):
                    logger.info(f"Pausing (battery: {battery.level}%, temp: {battery.temperature}°C)")
                    await asyncio.sleep(60)  # Wait 1 minute
                    continue

                # Process pending tasks
                await self._process_tasks()

                # Small delay
                await asyncio.sleep(5)

            except Exception as e:
                logger.error(f"Error in main loop: {e}", exc_info=True)
                await asyncio.sleep(10)

    def _get_battery_status(self) -> BatteryStatus:
        """Get current battery status"""
        try:
            battery = psutil.sensors_battery()

            if battery:
                return BatteryStatus(
                    level=int(battery.percent),
                    is_charging=battery.power_plugged,
                    temperature=35.0  # Placeholder - would use Android API
                )
        except Exception:
            pass

        # Default values if can't read
        return BatteryStatus(level=100, is_charging=True, temperature=30.0)

    def _should_process(self, battery: BatteryStatus) -> bool:
        """Determine if node should process tasks based on battery"""
        battery_config = self.config.get('battery', {})

        # Always process if charging
        if battery.is_charging:
            return True

        # Check minimum battery level
        min_level = battery_config.get('min_battery_level', 20)
        if battery.level < min_level:
            return False

        # Check temperature
        temp_limit = battery_config.get('temperature_limit', 40)
        if battery.temperature > temp_limit:
            logger.warning(f"Device too hot: {battery.temperature}°C")
            return False

        return True

    async def _process_tasks(self):
        """Process pending tasks"""
        # TODO: Connect to cluster coordinator
        # TODO: Fetch tasks from queue
        # TODO: Execute tasks based on mode
        # TODO: Return results

        # Placeholder
        logger.debug("Checking for tasks...")
        await asyncio.sleep(1)

    async def generate(self, prompt: str, mode: Optional[str] = None) -> str:
        """
        Generate text completion

        Args:
            prompt: Input prompt
            mode: Override operation mode ('local', 'cloud', or None for auto)

        Returns:
            Generated text
        """
        # Determine which backend to use
        use_mode = mode if mode else self._select_backend(prompt)

        if use_mode == 'local' and self.local_model:
            return await self._generate_local(prompt)
        elif use_mode == 'cloud' and self.cloud_client:
            return await self._generate_cloud(prompt)
        else:
            # Fallback
            if self.cloud_client:
                return await self._generate_cloud(prompt)
            elif self.local_model:
                return await self._generate_local(prompt)
            else:
                raise RuntimeError("No backend available")

    def _select_backend(self, prompt: str) -> str:
        """Smart backend selection for hybrid mode"""
        if self.mode == OperationMode.LOCAL:
            return 'local'
        elif self.mode == OperationMode.CLOUD:
            return 'cloud'

        # Hybrid mode - decide based on task complexity
        prompt_length = len(prompt.split())

        # Simple tasks → local
        if prompt_length < 50:
            return 'local' if self.local_model else 'cloud'

        # Complex tasks → cloud
        return 'cloud' if self.cloud_client else 'local'

    async def _generate_local(self, prompt: str) -> str:
        """Generate using local model"""
        logger.info("Generating with local model")

        # TODO: Implement actual llama.cpp call
        # This is a placeholder

        return f"[Local Model Response to: {prompt[:50]}...]"

    async def _generate_cloud(self, prompt: str) -> str:
        """Generate using cloud API"""
        logger.info(f"Generating with cloud ({self.cloud_client['provider']})")

        provider = self.cloud_client['provider']
        api_key = self.cloud_client['api_key']
        model = self.cloud_client['model']

        if provider == 'groq':
            return await self._call_groq(prompt, api_key, model)
        elif provider == 'openai':
            return await self._call_openai(prompt, api_key, model)
        else:
            raise ValueError(f"Unsupported provider: {provider}")

    async def _call_groq(self, prompt: str, api_key: str, model: str) -> str:
        """Call Groq API"""
        url = "https://api.groq.com/openai/v1/chat/completions"

        headers = {
            'Authorization': f'Bearer {api_key}',
            'Content-Type': 'application/json'
        }

        data = {
            'model': model,
            'messages': [{'role': 'user', 'content': prompt}],
            'max_tokens': 512
        }

        async with aiohttp.ClientSession() as session:
            async with session.post(url, headers=headers, json=data) as response:
                if response.status == 200:
                    result = await response.json()
                    return result['choices'][0]['message']['content']
                else:
                    error = await response.text()
                    raise RuntimeError(f"Groq API error: {error}")

    async def _call_openai(self, prompt: str, api_key: str, model: str) -> str:
        """Call OpenAI API"""
        base_url = self.cloud_client.get('base_url', 'https://api.openai.com/v1')
        url = f"{base_url}/chat/completions"

        headers = {
            'Authorization': f'Bearer {api_key}',
            'Content-Type': 'application/json'
        }

        data = {
            'model': model,
            'messages': [{'role': 'user', 'content': prompt}],
            'max_tokens': 512
        }

        async with aiohttp.ClientSession() as session:
            async with session.post(url, headers=headers, json=data) as response:
                if response.status == 200:
                    result = await response.json()
                    return result['choices'][0]['message']['content']
                else:
                    error = await response.text()
                    raise RuntimeError(f"OpenAI API error: {error}")


async def main():
    """Main entry point"""
    logger.info("="*60)
    logger.info("CH8 AGENT - ANDROID NODE")
    logger.info("="*60)

    # Create and start node
    node = AndroidNode()

    try:
        await node.start()
    except KeyboardInterrupt:
        logger.info("Received shutdown signal")
    finally:
        await node.stop()


if __name__ == '__main__':
    asyncio.run(main())
